refactored code by adding MVVM and clean architecture, also clean the code
