+++ BarcodeScanner/model/Error.swift
@@ -1,0 +1,13 @@
+//
+//  Error.swift
+//  BarcodeScanner
+//
+//  Created by farkhod on 22.01.2023.
+//
+
+import Foundation
+
+enum ScannerError: String {
+    case invalidDeviceError
+    case invalidScannedValue
+}
